# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Compte créé, session ouverte."
          },
          "400": {
            "description": "Les données envoyées sont invalides."
          },
          "409": {
            "description": "La requête entre en conflit avec l'état actuel de la ressource."
          },
          "429": {
            "description": "Trop de requêtes ont été envoyées dans un court instant."
          }
        },
        "summary": "Crée un compte et retourne un access token et un refresh token.",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Identifiants valides, session ouverte."
          },
          "400": {
            "description": "Les données envoyées sont invalides."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "429": {
            "description": "Trop de requêtes ont été envoyées dans un court instant."
          }
        },
        "summary": "Authentifie un utilisateur et retourne une nouvelle session.",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Nouveau couple access/refresh token."
          },
          "400": {
            "description": "Les données envoyées sont invalides."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "429": {
            "description": "Trop de requêtes ont été envoyées dans un court instant."
          }
        },
        "summary": "Fait tourner le refresh token et retourne un nouveau couple de tokens.",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Session révoquée."
          },
          "400": {
            "description": "Les données envoyées sont invalides."
          }
        },
        "summary": "Révoque le refresh token fourni (déconnexion de cet appareil).",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout-all": {
      "post": {
        "operationId": "AuthController_logoutAll",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Toutes les sessions ont été révoquées."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Révoque toutes les sessions actives de l'utilisateur courant.",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profil de l'utilisateur courant."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retourne le profil de l'utilisateur authentifié.",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/households/{householdId}/categories": {
      "get": {
        "operationId": "CategoriesController_list",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Catégories disponibles."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Liste les catégories (système et personnalisées) du household.",
        "tags": [
          "categories"
        ]
      },
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Catégorie créée."
          },
          "400": {
            "description": "Les données envoyées sont invalides."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          },
          "409": {
            "description": "La requête entre en conflit avec l'état actuel de la ressource."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crée une catégorie personnalisée (réservé à OWNER/ADMIN).",
        "tags": [
          "categories"
        ]
      }
    },
    "/households/{householdId}/categories/{categoryId}": {
      "patch": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Catégorie mise à jour."
          },
          "400": {
            "description": "Les données envoyées sont invalides."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          },
          "409": {
            "description": "La requête entre en conflit avec l'état actuel de la ressource."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Modifie une catégorie personnalisée (réservé à OWNER/ADMIN).",
        "tags": [
          "categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_remove",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Catégorie supprimée."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          },
          "409": {
            "description": "La requête entre en conflit avec l'état actuel de la ressource."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprime une catégorie personnalisée (réservé à OWNER/ADMIN). Les catégories système ne sont pas supprimables.",
        "tags": [
          "categories"
        ]
      }
    },
...
    "/households/{householdId}/exports/json": {
      "get": {
        "operationId": "ExportsController_exportJson",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection exportée en JSON."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Exporte la collection active du household au format JSON.",
        "tags": [
          "exports"
        ]
      }
    },
    "/households/{householdId}/exports/csv": {
      "get": {
        "operationId": "ExportsController_exportCsv",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection exportée en CSV."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Exporte la collection active du household au format CSV.",
        "tags": [
          "exports"
        ]
      }
    }
  },
  "info": {
    "title": "Notre Nid API",
    "description": "API de gestion de collection partagée pour Notre Nid.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "alex@example.com"
          },
          "password": {
            "type": "string",
            "example": "un-mot-de-passe-solide",
            "minLength": 8
          },
          "displayName": {
            "type": "string",
            "example": "Alex"
          }
        },
        "required": [
          "email",
          "password",
          "displayName"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "alex@example.com"
          },
          "password": {
            "type": "string"
          },
          "deviceName": {
            "type": "string",
            "example": "iPhone d'Alex"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "CategoryFieldSchemaDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "edition"
          },
          "label": {
            "type": "string",
            "example": "Édition"
          },
          "type": {
            "type": "string",
            "enum": [
              "string",
              "number",
              "boolean"
            ]
          },
          "required": {
            "type": "boolean"
          }
        },
        "required": [
          "key",
          "label",
          "type"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Vinyles"
          },
          "icon": {
            "type": "string",
            "example": "disc"
          },
          "metadataSchema": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryFieldSchemaDto"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Vinyles"
          },
          "icon": {
            "type": "string",
            "example": "disc"
          },
          "metadataSchema": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryFieldSchemaDto"
            }
          }
        }
      },
...
    }
  }
}